Fold Region/RegionControl into model.ts, delete modelControls.ts
modelControls.ts held only Region and RegionControl after the Minecraft
skin migration, and model.ts was its sole remaining importer (the
Control-union collapse already removed regionControls.tsx's need to
import from it directly). A single-type file with one consumer doesn't
earn its own module, so the types move into model.ts and the file goes
away.
